test(e2e): fold #434 schema-validation into existing restore specs
The restore_from_backup and restore_from_pv specs already stand up a source
cluster and a Backup / retained PV, so the standalone schema_validation spec
was re-creating the suite's most expensive fixtures (two extra clusters and a
backup) just to reach the same state.

Fold the two unique #434 assertions into those specs instead:
- record check: Backup.Status.SchemaVersion / PV annotation equals the source
  cluster's schema version;
- negative admission: a restore pinned to an older binary is rejected. The
  existing happy-path restore (default version >= schema) already covers admit.

Move the shared helpers into helpers_test.go and delete schema_validation_test.go.

Copilot-Session: b653a2f9-cc1a-400a-8e93-11232229023a
